+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## 3.14.0 
 * Add `PaymentMethodCustomerDataUpdated` webhook notification support
+* Add `phoneNumber` to `Address`
 
 ## 3.13.0
 * Make `junit-vintage` a test dependency (thanks @jamesbroadhead)